requirements.txt
numpy
soundfile
librosa
configparser
pandas
torch
onnx
onnxruntime-gpu
# nvidia-pyindex
# nvidia-tensorrt
# pycuda